update baru
- update tile
- update animasi
- update logika craft
- update logika destroy objek in game
- logika add prefab in game

+++ Assets/Script/Items/ItemBehavior.cs
@@ -6,7 +6,7 @@ public class ItemBehavior : MonoBehaviour
 {
     public Item item; // Referensi ke ScriptableObject Item
 
-    private int currentHealth;
+    private float currentHealth;
 
     void Start()
     {
